Moved the infrastructure code for distributed processes into Core projects
