+++ Core/Structures/Maybe.cs
@@ -1,4 +1,4 @@
-namespace HotelManagement.Core.Structures;
+namespace Core.Structures;
 
 public class Maybe<TSomething>
 {
