--- Sample/HotelManagement/HotelManagement.Tests/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace HotelManagement.Tests;
-
-public class Class1
-{
-}
