--- Sample/HotelManagement/HotelManagement/Core/Messaging/ICommandBus.cs
@@ -1,6 +1,0 @@
-namespace HotelManagement.Core.Messaging;
-
-public interface ICommandBus
-{
-    Task Send<TCommand>(TCommand command, CancellationToken ct);
-}
